requirements.txt
Flask==3.0.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
